requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
asyncer==0.0.2
attrs==23.1.0
bcrypt==4.0.1
boto3==1.28.35
botocore==1.31.35
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
coloredlogs==15.0.1
contourpy==1.1.0
cycler==0.11.0
dnspython==2.4.2
email-validator==2.0.0.post2
fastapi==0.103.0
ffmpy==0.3.1
filelock==3.12.2
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
gradio==3.41.2
gradio_client==0.5.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
imageio==2.31.2
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
lazy_loader==0.3
llvmlite==0.40.1
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numba==0.57.1
numpy==1.24.4
onnxruntime==1.15.1
opencv-python-headless==4.8.0.76
orjson==3.9.5
packaging==23.1
pandas==2.0.3
passlib==1.7.4
Pillow==10.0.0
platformdirs==3.10.0
pooch==1.7.0
protobuf==4.24.2
psycopg2==2.9.7
pydantic==2.3.0
pydantic_core==2.6.3
pydub==0.25.1
PyJWT==2.8.0
PyMatting==1.1.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.30.2
rembg==2.0.50
requests==2.31.0
rpds-py==0.9.2
s3transfer==0.6.2
scikit-image==0.21.0
scipy==1.11.2
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
sympy==1.12
tifffile==2023.8.25
toolz==0.12.0
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
watchdog==3.0.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2